Add files via upload
Using Online book store project the user can purchase books online instead of going out to a book store. The purpose of the project is to make a full functional online book store system that allow its users to search and purchase a book online based on category, author and subject. The selected books will be displayed in tabular form and then user can pick and order one or more books through COD or online payment.
1.1	Aim
The aim of this project is to develop software that covers all the aspects of management and operation of Book Store.

1.2	About Project
  An online book store software project that acts as a central                                                                        database containing various books in stock along with their title, author and cost. This project is a website that acts as a central bookstore. This web project will be developed using html JS CSS as the front end and java as a back-end. The SQL database stores various book related details. A user visiting the website can see a wide range of books arranged in respective categories. The user may select the desired book and view its price. The user may even search for specific books on the website. Once the user selects a book , he then has to fill in a form and the book is booked for the user.